+++ audionimbus/CHANGELOG.md
@@ -87,6 +87,7 @@
 - Implement `PartialEq`, `Eq` and `Hash` for `Scene`, `Simulator`, `Context`, `EmbreeDevice`, `OpenClDevice`, `OpenClDeviceList`, `RadeonRaysDevice`, `TrueAudioNextDevice`, `ImpulseResponse`, `SerializedObject`, `PanningEffect`, `PathEffect`, `BinauralEffect`, `ReflectionEffect`, `ReflectionMixer`, `VirtualSurroundEffect`, `DirectEffect`, `AmbisonicsPanningEffect`, `AmbisonicsBinauralEffect`, `AmbisonicsEncodeEffect`, `AmbisonicsDecodeEffect`, `AmbisonicsRotationEffect`, `Hrtf`, `EnergyField`, `ProbeArray`, `ProbeBatch`, `Reconstructor`, `InstancedMesh`, `StaticMesh`.
 - Implement `Clone` for `HrtfSettings`, `Sofa`.
 - add `AudioSettings` getters to `Simulator` and `Simulation`.
+- Implement `Default` for `DirectSimulationSettings`, `ConvolutionSettings` and `PathingSimulationSettings`.
 
 ### Removed